+++ pytsite/content/__init__.py
@@ -34,6 +34,7 @@ def __init():
     router.add_rule('/content/view/<string:model>/<string:id>', 'pytsite.content.ep.view')
     router.add_rule('/content/count/<string:model>/<string:id>', 'pytsite.content.ep.view_count')
     router.add_rule('/content/search/<string:model>', 'pytsite.content.ep.search', call='pytsite.content.ep.index')
+    router.add_rule('/content/ajax_search/<string:model>', 'pytsite.content.ep.ajax_search')
 
     # Propose route
     router.add_rule('/content/propose/<string:model>', 'pytsite.content.ep.propose',