+++ pytsite/metatag/__init__.py
@@ -1,11 +1,17 @@
 """PytSite MetaTag Module.
 """
 from ._functions import dump, dump_all, get, reset, t_set
-from pytsite import lang as _lang
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
-_lang.register_package(__name__)
+def __init():
+    from pytsite import lang, events
+
+    lang.register_package(__name__)
+    events.listen('pytsite.router.dispatch', reset)
+
+
+__init()